MonitoringServer/src/modules/predictive_analysis/requirements.txt
# Basic requirements
python-dotenv>=0.19.0
flask>=2.0.0
requests>=2.25.0

# Data analysis
numpy>=1.19.0
scikit-learn>=0.24.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter with binary package for easier installation 